Refactor Centralize session and default description

+++ viewModel/src/main/java/com/amsterdam/viewmodel/movieDetails/MovieDetailsInteractionListener.kt
@@ -11,7 +11,7 @@ interface MovieDetailsInteractionListener {
     fun onClickAddToList()
 
     fun onSaveMovieToList(
-        movieId: Int,
+        movieId: Long,
         listId: Long,
     )
